Implement functionality
The whole functionality of this tool is encapsulated in the Pons class
and accessible over the classes' static methods. Basically, the input
string is parsed to an array of adb shell commands and then executed one
after in another in a subshell.

It might be faster to concatenate all adb shell commands with a
semicolon, but there is an upper limit on the service input length.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>