Add home button signal option
In order to unlock the screen if the phone was turned off I have added a
home button signal option. For the option parsing I am using trollop. It
can be improved by reading additional information like the power state
of the display. This way the use of the home button signal could be
automatized.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>